adjust retirement income for inflation. closes #15

+++ src/web/mod.rs
@@ -39,7 +39,7 @@ impl Into<simulator::Vars> for Params {
             tax_strategy: self.tax_strategy.unwrap_or(String::from("")),
             with_fiscal_partner: self.fiscal_partner.unwrap_or(false),
             with_heffingskorting: self.heffingskorting.unwrap_or(false),
-            aow_amount: self.aow_amount.unwrap_or(0.00) as u64,
+            aow_amount: self.aow_amount.unwrap_or(0.00),
             aow_starts_after_x_years: match self.aow_start_year {
                 None => 0,
                 Some(y) => {
